requirements.txt
google-generativeai==0.8.5
python-dotenv==1.1.1
requests==2.32.5
rich==14.1.0
typer==0.16.1
